radonc: mirror upstream radon structurally (files, names, tests)
Restructured the CC crate to mirror the original radon Python library so
it can be diffed against upstream and updated when radon changes:

- visitors.rs mirrors radon/visitors.py: ComplexityVisitor with the same
  public surface (from_code/from_ast, complexity, functions_complexity,
  classes_complexity, total_complexity, blocks), Function/Class block
  types with the same fields (name/lineno/col_offset/endline/
  is_method/classname/closures), GET_COMPLEXITY, code2ast. Counting
  matches radon's generic_visit: Try handlers+else, BoolOp values-1,
  If/IfExp +1, Match minus underscore, For/While+else, comprehension
  ifs+1, Assert+1 non-recursive, lambda +0 body walked, and visit_FunctionDef
  / visit_ClassDef scope rules (class body = 1 + methods' CCs).
- complexity.rs mirrors radon/complexity.py: cc_rank (A-F), average_complexity,
  sorted_results (score/lines/alpha), add_inner_blocks, cc_visit, cc_visit_ast.
- tests/test_complexity_visitor.rs mirrors the upstream test data:
  SIMPLE_BLOCKS, SINGLE_FUNCTIONS, FUNCTIONS, CLASSES, MODULE (GENERAL_CASES),
  CLOSURES, CONTAINERS case tables with radon's exact expected values.

The port caught two real deviations while mirroring the tests: lambda was
counted +1 (radon: +0, body walked) and the source-order walker entry was
skipping the counting override at module level. Fixed; parity is now
pinned by the mirrored tables (6 radonc tests) on top of the 141 scan-core
tests.

PRD R26 recorded: radonc must maintain structural fidelity to upstream
radon so updates and handover-to-maintainers stay mechanical.

+++ scanner/src/main.rs
@@ -154,7 +154,7 @@ impl<'a> SourceOrderVisitor<'a> for ScanState<'a> {
                 // The radon-equivalent CC comes from the radonc crate (its rules
                 // match radon 6.0.1 exactly; the visitor still tracks decisions
                 // for the closures/latent-class gate).
-                let cc = radonc::function_cc(f);
+                let cc = radonc::function_cc(f) as u32;
                 // repo-wide collections: module-level defs + duplicate
                 // candidates (the def line is the NAME's line — ruff's
                 // FunctionDef range starts at the decorator)
